+++ crates/ruma-client-api/CHANGELOG.md
@@ -14,6 +14,7 @@ Improvements:
 * Add support for querying relating events (MSC2675)
 * Move `filter::RelationType` to `ruma_common::events::relations`
 * Add unstable support for discovering an OpenID Connect server (MSC2965)
+* Add `SpaceRoomJoinRule::KnockRestricted` (MSC3787)
 
 # 0.14.1